x2t - конвертация ms<->open office. готово.
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@63021 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeOdfFile/src/docx/oox_chart_axis.cpp
@@ -20,7 +20,7 @@ _CP_PTR(oox_axis_content) oox_axis_content::create(int type)
 oox_axis_content::oox_axis_content(int type/*,std::wstring name*/)
 {
     //id_ = abs((int)*((_UINT32*)this));
-    id_ = abs((int)this);
+    id_ = abs((long)this);
 	type_=type;		//dimension
 }
 void oox_axis_content::oox_serialize(std::wostream & _Wostream)
